register routes broken out to register.js
